#1 Basics of SpaceGame
Initialized the use of Player controller to control the movement of the spaceship that the user has control of.

Initialized the randomized spawning of enemies, which can have determined "Speed" as well as locations of spawning for the spaceships.

Player Ship has as possibility to shoot bullets, which can be modified by:
Speed
Life Time of the bullet
Size of the bullet (Only inside code itself)

-- Bullet has currently no other function other than shooting and clearing the bullet, needs a collision detection consider the use of TriggerBoxes to apply damage to Enemy ships or other elements.

Enemy Ship Spawners can also be used for other objects such as Meteorites or elements that fly around in space like rocks.

Notes :
Currently no Audio
Health System is missing
Bullet working is not available
No Main Menu / Pause Menu